Relocate activation functions from layers.py to separate .py
